pyproject.toml
[tool.poetry]
name = "cloud-utils"
version = "0.0.0"
description = "A single interface for interacting with multiple cloud platforms."
authors = ["Matt Hayward <mhayward122@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.34.20"
google-cloud-compute = "^1.15.0"
google-cloud-dns = "^0.35.0"


[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.devmagic.cloud/api/v4/projects/6/packages/pypi"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
